Refactor local progress repository
